+++ FinancialFreedom/app/src/main/java/com/example/financialfreedom/activity/MyOptionalActivity.kt
@@ -179,8 +179,8 @@ class MyOptionalActivity : BaseActivity() {
                      * 通过MyOptionalActivity的静态对象调用相关的方法
                      */
                     val intent = Intent(myOptionalActivity, MyOptionalDetailActivity::class.java)
-                    /* 向活动传递数据：注意为Int型 */
-                    intent.putExtra("stock_code", stockCode.toInt())
+                    /* 向活动传递数据：注意为确保"0"开头的如深市代码能正确传递，均使用String类型进行传递 */
+                    intent.putExtra("stock_code", stockCode)
                     myOptionalActivity.startActivity(intent) //开启活动
                     Log.d("MyOptionalActivity", "click stockCode:$stockCode")
                 }
